migrate homepage content to be loading page & migrate index content to about page
