+++ cypress/integration/chapter/chapter.spec.js
@@ -5,7 +5,7 @@ describe('Chapter view', () => {
     });
 
     it('should display chapter details', () => {
-        cy.visit('/home');
+        cy.visit('/');
         cy.get(':nth-child(1) > .card > .card-body > .card-title a').click();
         cy.get('#chapter > h4').should('exist').should('be.visible');
     });
