+++ frontend/app/components/main-header.hbs
@@ -24,38 +24,26 @@
 
   <div class="collapse navbar-collapse" id="mainTopNav">
     <ul class="navbar-nav ml-auto text-{{this.config.content.APP.header_text}} ">
-      {{#if this.me.user.id}}
 
-        <li class="nav-item text-{{this.config.content.APP.header_text}}">
+      <li class="nav-item text-{{this.config.content.APP.header_text}}">
 
-          <LinkTo @route="home" class="nav-link" data-toggle="collapse" data-target="#mainTopNav.show">
-            {{t "buttons.learn"}}
-          </LinkTo>
-        </li>
-
+        <LinkTo @route="index" class="nav-link" data-toggle="collapse" data-target="#mainTopNav.show">
+          {{t "buttons.learn"}}
+        </LinkTo>
+      </li>
+      {{#if this.me.isAuthenticated}}
         <li class="nav-item text-{{this.config.content.APP.header_text}}">
           <LinkTo @route="teach" class="nav-link" data-toggle="collapse" data-target="#mainTopNav.show">
             {{t "buttons.teach"}}
           </LinkTo>
         </li>
-        <li class="nav-item text-{{this.config.content.APP.header_text}}">
-          <LinkTo @route="index" class="nav-link" data-toggle="collapse" data-target="#mainTopNav.show">
-            {{t "buttons.about"}}
-          </LinkTo>
-        </li>
-      {{else}}
-        <li class="nav-item text-{{this.config.content.APP.header_text}}">
-          <LinkTo @route="home" class="nav-link" data-toggle="collapse" data-target="#mainTopNav.show">
-            {{t "buttons.learn"}}
-          </LinkTo>
-        </li>
-        <li class="nav-item text-{{this.config.content.APP.header_text}}">
-          <a class="nav-link" href="https://wikonnect.org">
-            {{t "buttons.about"}}
-          </a>
-
-        </li>
       {{/if}}
+      <li class="nav-item text-{{this.config.content.APP.header_text}}">
+        <LinkTo @route="about" class="nav-link" data-toggle="collapse" data-target="#mainTopNav.show">
+          {{t "buttons.about"}}
+        </LinkTo>
+      </li>
+
     </ul>
     <form class="form-inline ml-lg-5 search-form" {{on "submit" this.search}}>
       <Input @id="search" @class="form-control mr-sm-2 search-input" @type="text" @value={{this.search_term}} />
