+++ frontend/app/routes/signup.js
@@ -6,7 +6,7 @@ export default class SignupRoute extends Route {
 
   beforeModel() {
     if (this.me.user) {
-      this.transitionTo('home');
+      this.transitionTo('index');
     }
   }
