+++ frontend/app/templates/about.hbs
@@ -1,0 +1,28 @@
+<div id="container">
+
+
+  <div class=" w-100 m-0 p-0 bg-orange kkl-cover d-flex align-items-end">
+
+    <div class="display-4 bg-orange kkl-jumbo-text p-4">Welcome to Keep Kenya Learning</div>
+  </div>
+  <div class="row p-5 m-0">
+
+
+    <div class="col-sm">
+      <h4 class="text-primary">Why we exist</h4>
+      <p>
+        We support Kenyan families and teachers who want to provide learning opportunities for children whether schools
+        are open or closed.
+      </p>
+    </div>
+    <div class="col-sm">
+      <h4 class="text-primary">What we offer</h4>
+      <p>
+        Find fun learning activities aligned with 8-4-4 and the new Competency Based Curriculum (CBC). Update your
+        profile
+        to
+        see activities for your children’s ages and interests.
+      </p>
+    </div>
+  </div>
+</div>
