+++ frontend/app/templates/index.hbs
@@ -1,28 +1,20 @@
-<div id="container">
-
-
-  <div class=" w-100 m-0 p-0 bg-orange kkl-cover d-flex align-items-end">
-
-    <div class="display-4 bg-orange kkl-jumbo-text p-4">Welcome to Keep Kenya Learning</div>
+<div class="chapter-filters-section">
+  <div class="container pb-3 pt-3">
+    <ChapterTags::TagsList @tagsList={{this.tagsList}} @toggleTagSelection={{this.toggleTagSelection}}
+                           @selectedFilterTags={{this.selectedFilterTags}} @clearAllTagFilters={{this.clearAllTagFilters}} />
   </div>
-  <div class="row p-5 m-0">
+</div>
 
-
-    <div class="col-sm">
-      <h4 class="text-primary">Why we exist</h4>
-      <p>
-        We support Kenyan families and teachers who want to provide learning opportunities for children whether schools
-        are open or closed.
-      </p>
-    </div>
-    <div class="col-sm">
-      <h4 class="text-primary">What we offer</h4>
-      <p>
-        Find fun learning activities aligned with 8-4-4 and the new Competency Based Curriculum (CBC). Update your
-        profile
-        to
-        see activities for your children’s ages and interests.
-      </p>
+<div class="container">
+  <div class="row mt-5 ">
+    <Home::Chapters @chapters={{this.chapters}}>
+    </Home::Chapters>
+  </div>
+  <div class="row">
+    <div class="col">
+      <InfinityLoader @infinityModel={{this.chapters}} @loadingText="Loading more awesome records..."
+                      @loadedText="Loaded all the records!" @triggerOffset={{500}} @eventDebounce={{1000}}/>
     </div>
   </div>
+
 </div>
